Sorting demonstration
Each sorting algorithm is coded as a type with a run() receiver function
which performs the sort. The type also contains references to two
channels for the purpose of reporting element pair comparisons and
element pair swaps.

The channels are read by processors which will report activity and
progress to the user. When the run() function of a sort completes, a
special event is sent to the channel to signal completed processing.
Otherwise normal events are reported with relevent details. The channel
processor routines are also connected to a master supervisor channel
which is monitored. The main goroutine is blocked from exiting until all
sorting event channel processors have reported completion.

To allow starting each sorting algorithm on matching unsorted arrays,
the elements of an initial array are copied into a data member in of the
constructor.